victara: cmactions: Don't keep the screen on on the lock screen
Only start using the IR sensors to keep the screen on for the
User-Aware Display mode when the phone is unlocked.

Prior to this change it was far too easy to accidentally trigger
the IR sensors while plugging your phone in which could leave the
display on indefinitely if your USB chargers are too close to
your house plants.  For example.

Change-Id: I5fbd82a9679fa9d7711cff18b41435309f6abde8