victara: cmaction: Honour the default selection of a camera
When the user has selected a default camera for the still shot
intent, honour it both in the secure and normal cases.

If the user hasn't yet selected a default camera and we are
lanching the normal (not secure) camera, allow them to use
the activity picker to pick the one that they want to use.

Change-Id: I7b0f1dfad27b415151b63f0308cf44849e985c58